[Appearance] factor out ThemeSelector and use it here
