+++ ui/page/settings/index.js
@@ -5,7 +5,6 @@ import {
   doSetDaemonSetting,
   doClearDaemonSetting,
   doSetClientSetting,
-  doSetDarkTime,
   doEnterSettingsPage,
   doExitSettingsPage,
 } from 'redux/actions/settings';
@@ -20,16 +19,12 @@ const select = (state) => ({
   allowAnalytics: selectAllowAnalytics(state),
   isAuthenticated: selectUserVerifiedEmail(state),
   showNsfw: selectShowMatureContent(state),
-  currentTheme: makeSelectClientSetting(SETTINGS.THEME)(state),
-  themes: makeSelectClientSetting(SETTINGS.THEMES)(state),
-  automaticDarkModeEnabled: makeSelectClientSetting(SETTINGS.AUTOMATIC_DARK_MODE_ENABLED)(state),
   clock24h: makeSelectClientSetting(SETTINGS.CLOCK_24H)(state),
   autoplay: makeSelectClientSetting(SETTINGS.AUTOPLAY)(state),
   autoDownload: makeSelectClientSetting(SETTINGS.AUTO_DOWNLOAD)(state),
   hideBalance: makeSelectClientSetting(SETTINGS.HIDE_BALANCE)(state),
   floatingPlayer: makeSelectClientSetting(SETTINGS.FLOATING_PLAYER)(state),
   hideReposts: makeSelectClientSetting(SETTINGS.HIDE_REPOSTS)(state),
-  darkModeTimes: makeSelectClientSetting(SETTINGS.DARK_MODE_TIMES)(state),
   myChannelUrls: selectMyChannelUrls(state),
   user: selectUser(state),
 });
@@ -40,7 +35,6 @@ const perform = (dispatch) => ({
   toggle3PAnalytics: (allow) => dispatch(doToggle3PAnalytics(allow)),
   setClientSetting: (key, value) => dispatch(doSetClientSetting(key, value)),
   clearPlayingUri: () => dispatch(doSetPlayingUri({ uri: null })),
-  setDarkTime: (time, options) => dispatch(doSetDarkTime(time, options)),
   openModal: (id, params) => dispatch(doOpenModal(id, params)),
   enterSettings: () => dispatch(doEnterSettingsPage()),
   exitSettings: () => dispatch(doExitSettingsPage()),
